+++ x/staking/keeper/delegation_test.go
@@ -286,7 +286,6 @@ func (s *KeeperTestSuite) TestUnbondingDelegation() {
 		0,
 		time.Unix(0, 0).UTC(),
 		math.NewInt(5),
-		0,
 		address.NewBech32Codec("cosmosvaloper"), address.NewBech32Codec("cosmos"),
 	)
 
@@ -343,8 +342,8 @@ func (s *KeeperTestSuite) TestUnbondingDelegationsFromValidator() {
 		0,
 		time.Unix(0, 0).UTC(),
 		math.NewInt(5),
-		0,
-		address.NewBech32Codec("cosmosvaloper"), address.NewBech32Codec("cosmos"),
+		address.NewBech32Codec("cosmosvaloper"),
+		address.NewBech32Codec("cosmos"),
 	)
 
 	// set and retrieve a record
@@ -716,7 +715,7 @@ func (s *KeeperTestSuite) TestGetRedelegationsFromSrcValidator() {
 
 	rd := stakingtypes.NewRedelegation(addrDels[0], addrVals[0], addrVals[1], 0,
 		time.Unix(0, 0), math.NewInt(5),
-		math.LegacyNewDec(5), 0, address.NewBech32Codec("cosmosvaloper"), address.NewBech32Codec("cosmos"))
+		math.LegacyNewDec(5), address.NewBech32Codec("cosmosvaloper"), address.NewBech32Codec("cosmos"))
 
 	// set and retrieve a record
 	err := keeper.SetRedelegation(ctx, rd)
@@ -746,7 +745,7 @@ func (s *KeeperTestSuite) TestRedelegation() {
 
 	rd := stakingtypes.NewRedelegation(addrDels[0], addrVals[0], addrVals[1], 0,
 		time.Unix(0, 0).UTC(), math.NewInt(5),
-		math.LegacyNewDec(5), 0, address.NewBech32Codec("cosmosvaloper"), address.NewBech32Codec("cosmos"))
+		math.LegacyNewDec(5), address.NewBech32Codec("cosmosvaloper"), address.NewBech32Codec("cosmos"))
 
 	// test shouldn't have and redelegations
 	has, err := keeper.HasReceivingRedelegation(ctx, addrDels[0], addrVals[1])
@@ -1107,14 +1106,14 @@ func (s *KeeperTestSuite) TestUnbondingDelegationAddEntry() {
 		creationHeight,
 		time.Unix(0, 0).UTC(),
 		math.NewInt(10),
-		0,
-		address.NewBech32Codec("cosmosvaloper"), address.NewBech32Codec("cosmos"),
+		address.NewBech32Codec("cosmosvaloper"),
+		address.NewBech32Codec("cosmos"),
 	)
 	var initialEntries []stakingtypes.UnbondingDelegationEntry
 	initialEntries = append(initialEntries, ubd.Entries...)
 	require.Len(initialEntries, 1)
 
-	isNew := ubd.AddEntry(creationHeight, time.Unix(0, 0).UTC(), math.NewInt(5), 1)
+	isNew := ubd.AddEntry(creationHeight, time.Unix(0, 0).UTC(), math.NewInt(5))
 	require.False(isNew)
 	require.Len(ubd.Entries, 1) // entry was merged
 	require.NotEqual(initialEntries, ubd.Entries)
@@ -1123,7 +1122,7 @@ func (s *KeeperTestSuite) TestUnbondingDelegationAddEntry() {
 	require.Equal(ubd.Entries[0].Balance, math.NewInt(15))                   // 10 from previous + 5 from merged
 
 	newCreationHeight := int64(11)
-	isNew = ubd.AddEntry(newCreationHeight, time.Unix(1, 0).UTC(), math.NewInt(5), 2)
+	isNew = ubd.AddEntry(newCreationHeight, time.Unix(1, 0).UTC(), math.NewInt(5))
 	require.True(isNew)
 	require.Len(ubd.Entries, 2) // entry was appended
 	require.NotEqual(initialEntries, ubd.Entries)
@@ -1131,7 +1130,6 @@ func (s *KeeperTestSuite) TestUnbondingDelegationAddEntry() {
 	require.Equal(newCreationHeight, ubd.Entries[1].CreationHeight)
 	require.Equal(ubd.Entries[0].Balance, math.NewInt(15))
 	require.Equal(ubd.Entries[1].Balance, math.NewInt(5))
-	require.NotEqual(ubd.Entries[0].UnbondingId, ubd.Entries[1].UnbondingId) // appended entry has a new unbondingID
 }
 
 func (s *KeeperTestSuite) TestSetUnbondingDelegationEntry() {
@@ -1149,8 +1147,8 @@ func (s *KeeperTestSuite) TestSetUnbondingDelegationEntry() {
 		creationHeight,
 		time.Unix(0, 0).UTC(),
 		math.NewInt(5),
-		0,
-		address.NewBech32Codec("cosmosvaloper"), address.NewBech32Codec("cosmos"),
+		address.NewBech32Codec("cosmosvaloper"),
+		address.NewBech32Codec("cosmos"),
 	)
 
 	// set and retrieve a record
@@ -1180,8 +1178,7 @@ func (s *KeeperTestSuite) TestSetUnbondingDelegationEntry() {
 	require.Len(resUnbonding.Entries, 1)
 	require.NotEqual(initialEntries, resUnbonding.Entries)
 	require.Equal(creationHeight, resUnbonding.Entries[0].CreationHeight)
-	require.Equal(initialEntries[0].UnbondingId, resUnbonding.Entries[0].UnbondingId) // initial unbondingID remains unchanged
-	require.Equal(resUnbonding.Entries[0].Balance, math.NewInt(10))                   // 5 from previous entry + 5 from merged entry
+	require.Equal(resUnbonding.Entries[0].Balance, math.NewInt(10)) // 5 from previous entry + 5 from merged entry
 
 	// set unbonding delegation entry for newCreationHeight
 	// new entry is expected to be appended to the existing entries
@@ -1202,11 +1199,6 @@ func (s *KeeperTestSuite) TestSetUnbondingDelegationEntry() {
 	require.NotEqual(resUnbonding.Entries[0], resUnbonding.Entries[1])
 	require.Equal(creationHeight, resUnbonding.Entries[0].CreationHeight)
 	require.Equal(newCreationHeight, resUnbonding.Entries[1].CreationHeight)
-
-	// unbondingID is incremented on every call to SetUnbondingDelegationEntry
-	// unbondingID == 1 was skipped because the entry was merged with the existing entry with unbondingID == 0
-	// unbondingID comes from a global counter -> gaps in unbondingIDs are OK as long as every unbondingID is unique
-	require.Equal(uint64(2), resUnbonding.Entries[1].UnbondingId)
 }
 
 func (s *KeeperTestSuite) TestUndelegateWithDustShare() {