+++ x/staking/keeper/val_state_change.go
@@ -456,19 +456,12 @@ func (k Keeper) BeginUnbondingValidator(ctx context.Context, validator types.Val
 		return validator, fmt.Errorf("should not already be unbonded or unbonding, validator: %v", validator)
 	}
 
-	id, err := k.IncrementUnbondingID(ctx)
-	if err != nil {
-		return validator, err
-	}
-
 	validator = validator.UpdateStatus(types.Unbonding)
 
 	headerInfo := k.HeaderService.HeaderInfo(ctx)
 	// set the unbonding completion time and completion height appropriately
 	validator.UnbondingTime = headerInfo.Time.Add(params.UnbondingTime)
 	validator.UnbondingHeight = headerInfo.Height
-
-	validator.UnbondingIds = append(validator.UnbondingIds, id)
 
 	// save the now unbonded validator record and power index
 	if err = k.SetValidator(ctx, validator); err != nil {
@@ -496,14 +489,6 @@ func (k Keeper) BeginUnbondingValidator(ctx context.Context, validator types.Val
 	}
 
 	if err := k.Hooks().AfterValidatorBeginUnbonding(ctx, consAddr, str); err != nil {
-		return validator, err
-	}
-
-	if err := k.SetValidatorByUnbondingID(ctx, validator, id); err != nil {
-		return validator, err
-	}
-
-	if err := k.Hooks().AfterUnbondingInitiated(ctx, id); err != nil {
 		return validator, err
 	}